added global counters for file storage sizes and files

+++ services/mass_storage_service.h
@@ -40,7 +40,7 @@
 
 struct _MS_data {
     FATFS test;
-    uint16_t stores_fsize[3];   /*total file size in each store, fotos, su_log, ev_log */
+    uint32_t stores_fsize[3];   /*total file size in each store, fotos, su_log, ev_log */
     uint16_t stores_fcount[3];  /*file count in each store, fotos, su_log, ev_log */
     uint32_t ev_temp_log;
 };
@@ -54,7 +54,6 @@ extern void cnv8_32(uint8_t *from, uint32_t *to);
 extern void cnv32_8(const uint32_t from, uint8_t *to);
 
 //ToDo
-//  add global counters for file and size, add check for array limits.
 //	check type casting for snprintf & %d conversions
 //	error checking for sprintf
 //  strtol proper checking, return value range is ok
@@ -66,6 +65,7 @@ extern void cnv32_8(const uint32_t from, uint8_t *to);
 //  maybe file paths should be const variable instead of definitions.
 
 //Finished
+//  add global counters for file and size, add check for array limits.
 //  add check for MAX_FILE for loop, hard limit.
 //  check if sprintf is ok for the job, used snprintf
